add left eye
need to fix right eye, something is wrong with it still. if i go super far away from it, the box is lower than it should be

left eye is too high and too the right, gets worse the further away i go. i think i'm interpreting the quaternions incorrectly?